refactor(Problem2): 클래스 분리 및 메소드명 변경
