+++ src/main/java/onboarding/problem2/Constant.java
@@ -1,0 +1,5 @@
+package onboarding.problem2;
+
+public class Constant {
+    static final String EXCEPTION = "[ERROR]";
+}
